Enable indexless (headless) deploy
Moves all of the Elasticsearch configuration to a
docker-compose.override.yml file. This file will be used by default when
no config file is specified and will override the defaults in the
docker-compose.yml file. If the latter file is explicitly referenced
(`-f docker-compose.yml`) then only that config will be used and no
Elasticsearch container will be built and search will be disabled in
Archivematica. To see the config that will be used run in either case, run
`docker-compose -f docker-compose.yml config` or
`docker-compose config`.