Add elasticsearch virtual memory setting to docs
This commit adds a vm.max_map_count setting and troubleshooting
section to README.md.

A small correction has been made to a link to the Compose 1.7
branch. This has been turned into a Github permalink.

The opportunity has been taken to refactor the entire README so that
it is in-line with the Archivematica documentation style-guide.

* https://github.com/artefactual/archivematica-docs/wiki/Style-guide#formatting-and-punctuation

Lines have been truncated and links, where appropriate, have been
converted to markdown- reference-style.