Add helpers for AMAUAT functions
Make it as easy as possible for developers to run automated tests via
this Makefile.

We add a helper for black-box tests. Black-box tests are a minimum
that new features being developed in Archivematica should pass
against.

We also provide a listing output mechanism. This recipe lists all of
the tests that we expect to be able to run via Docker without any
additional configuration.