Create complete recipes for pip-compile
This commit introduces a storage service command for pip-compile.
For both the Storage Service and Archivematica, 'clean' and compile
'all' is run so that they are compiled fresh every run. This costs a
little extra time but ensures we're always up-to-date with what
pip-tools tells us. Ownership of requirements output from Docker are
set to UID:GID of the caller.

Co-authored-by: Jesús García Crespo <jesus@sevein.com>