Add MySQL config file to override default settings
The new `am-tuning.cnf` file overrides some MySQL settings that are often too
low in default deployments.

The following variables have been added to the file:

* max_allowed_packet = 1G
* innodb_log_file_size = 1G

Mounting the `am-tuning.cnf` as a new volume with a relative path on the docker
host allows to apply changes in MySQL easily. The user only needs to change
that file in the docker host, and then restart or rebuild the container.