Moves system level chown to containers
Additionally we add the clean commands to the compile call which was
this author's original intention but didn't realise all didn't mean
everything. Compile commands will always clean first, and so will
always build. This maximizes the chances of the developer updating
requirements as needed without other changes being missed.